backend/requirements.txt
flask>=2.2.3
flask-cors>=3.0.10
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.80.0
pyopenssl>=23.0.0
werkzeug>=2.2.0